requirements.txt
ChatterBot==1.0.2
discord.py==1.7.3
discord==1.7.3
python-dotenv==0.18.0
